+++ cypress/integration/configuration/channels.js
@@ -14,12 +14,13 @@ import { HEADER_SELECTORS } from "../../elements/header/header-selectors";
 import { DRAFT_ORDER_SELECTORS } from "../../elements/orders/draft-order-selectors";
 import { ORDERS_SELECTORS } from "../../elements/orders/orders-selectors";
 import { BUTTON_SELECTORS } from "../../elements/shared/button-selectors";
+import { SHARED_ELEMENTS } from "../../elements/shared/sharedElements";
 import { createChannelByView } from "../../steps/channelsSteps";
 import { urlList } from "../../url/urlList";
 import { deleteChannelsStartsWith } from "../../utils/channelsUtils";
 
 describe("Channels", () => {
-  const channelStartsWith = "Cypress:";
+  const channelStartsWith = `CyChannels:`;
   const currency = "PLN";
 
   before(() => {
@@ -67,6 +68,10 @@ describe("Channels", () => {
     cy.addAliasToGraphRequest("InitialProductFilterAttributes");
     cy.visit(urlList.products);
     cy.wait("@InitialProductFilterAttributes");
+    cy.get(SHARED_ELEMENTS.progressBar)
+      .should("not.exist")
+      .get(PRODUCTS_LIST.emptyProductRow)
+      .should("not.exist");
     cy.get(PRODUCTS_LIST.productsList)
       .first()
       .click()