+++ src/utils/richText/misc.ts
@@ -1,0 +1,5 @@
+import { OutputData } from "@editorjs/editorjs";
+
+export const getParsedDataForJsonStringField = (
+  data: OutputData
+): string | null => (data?.blocks?.length ? JSON.stringify(data) : null);